Code revision:
Fixed changed CSS classes for the video cards.
Changed the logic to hide elements so that any element can be hidden, regardless of its type and its pre-existing CSS style (now any element that needs to be hidden is wrapped in a custom container which can be hidden without interfering with pre-existing CSS).