First commit of the browser_action:
Browser action includes a label of the current channel plus a button to add/remove that channel to the Spoiler-Free Mode channels. To achieve that the content_script sets a local storage item "currentChannel" as it determines the current channel. The browser action listens for storage changes and in return updates the current channel label and the button; Whether the button functions to add or to remove the channel depends on whether the channel is in the list of the option [sfmChannels]. Because of that, the browser_action needs to listen to changes to the option [sfmChannels] and update the button status accordingly;
To achieve this, the classes Platform and Channel were created;