Changes to the logic of sfmEnabledForPage and SFM on video lists;
If sfmEnabledForPage is UNDETERMINED, the page will be configure as if SFM was disabled; It will only be reconfigured if the channel becomes determined or an option changes that triggers sfmEnabledForPage to change or changes for which channels the SFM is enabled. On pages that are not channel pages, each video item is inspected for its channel and then configured according to whether the SFM is enabled for that specific channel (on channel (sub)pages all video list items are configured according to the channel rules);
Plus: Frozen additional objects: TwitchPageType, TWITCH_PLATFORM, ALL_PLATFORMS;
Plus: Twitch Channels names are now converted to lower case when building a channel (the name is always lowercase, the display name can be mixed case). Before comparison would fail if because it is case-sensitive;