imgs+tooltips for videoListItemToolbar, handling async loaded videos:
The buttons of a video list item toolbar now have proper images and tooltips that change their message depending on whether the hideable content is visible or hidden;
Plus: Now listening for added nodes on the div.tw-tower that is the parent node for all the video list items. When a user scrolls down the list and reaches the end, the next entries are automatically loaded. A MutationObserver is used to listen for added nodes and then trigger a reconfiguration of all video list items;