Changed SVG's color of video list item button images to grey:
Used the same color as the timestamp of videos so it blends in nicely. The grey is indepedent of whether the user uses the Twitch default theme (white background) or the BetterTTV dark theme (black background). The white color that was used before only worked well with the BetterTTV dark theme;