Refactoring/bug fixing for feature "Enabled SFM platform-dependent"
- The channel link anchor is actually not a anchor anymore when being
on the channel main page (it's still an anchor on any channel directory
or on a video page). The parsing is adjusted accordingly. Sadly, the
channel's qualified name is not part of this new channel link div, so
we try to take it from the (hopefully) already parsed channel from the
URL.
- Improved the English messages for the popup menu.
- Improved the HTML layout / CSS styling of the popup menu.
- Fixed a bug that wrote the string "null" to the channelDisplayName
data attribute of the channel element of the popup if the channel's
display name was null.