Fixed toolbars and hidden elements not cleaned up after page change
When watching a VOD where SFM is enabled, then switching on-page to a
live channel (e.g. watching twitch.tv/overwatchleague VOD, then switch
to the twitch.tv/emongg live stream by clicking on the link on the left
toolbar), the Open End Player Toolbar was removed and the hidden
elements were not made visible.
Now, the page is cleaned up after a page change.