Fixed popup not displaying channel if not options stored yet
When reading the necessary options in the popup, the key name of the
sfmEnabledChannels option was used as the default value. So if no
options were present yet (new installation), instead of an empty array
a string would be the default value, leading to an Error which would
prevent the channel div to become visible.