Fixed on-page ch changing lead to previous ch determined as current ch
Because the channel link anchor is not changed immediately but async,
that update sometimes happens after the parsing of the channel after a
page change. In this case, the old channel was determined as the current
channel instead of the new channel.
A MutationObserver now observes the href attribute of the channel link
anchor element and updates the channel to null (needs re-determination)
when the href changes.