Fix async loaded video list items not configured
The check whether the configuration of async loaded video list items is
completed is now more robust (when you scroll down a video list, more videos are loaded
asynchronously).
Instead of just checking whether titles/previews/durations exist
(which is always the case after the first video list items are loaded),
it is now checked whether the number of found video list items is equal
to the number of found titles/previews/durations.
This check is necessary because new video list items are first displayed
with placeholder boxes instead of actual information.