Fix: Async loaded videos were not configured after the page cfg timeout
If the page configuration timeout (common.js:PAGE_CONFIGURATION_TIMEOUT)
has been reached, the async loaded video list items were not configured
correctly if loading took some time.
The reason for this was that when the added items were detected the
re-configuration was triggered once. But if it failed to configure the
new items then because they were not fully loaded yet, it would not try
to configure them again. So now we reset the page content changed time
whenever items are added so that the configure job tries to configure
the page again (until the timeout is reached again).