Fix: Infinite loop when async videos added
When videos were added asynchronously to video lists, many of them are wrapped
in an opnd-container to hide them. This was observed by the observer as
well and another addition event was triggered which lead to the whole
process beginning again and never ending. So now we filter the added
video list items: Items that have the CSS class opnd-container are not
counted as added items (as this is just a wrapping process).