Fixed progress visible after Twitch commercials 🐛
When not using Adblock Plus, Twitch plays commercials every now and then. After
the commercial ends and the video playback continues, it seems that Twitch
manipulates the DOM of the progress elements in a way that removes the CSS
classes added by Open End to hide them.
To fix this we now flag the parent player element when the Spoiler-Free Mode is
active. The player element does not get modified so this works regardless of
what Twitch does to the elements inside the player.
Plus, some minor code clean up.